Day 16 part 2 solved, but there is room for efficiency improvements.
The backward fft algorithm can be rewritten to accept the number of
phases to simulate. Instead of just summing, it can multiply each number
in the backward series by the number of phases minus the offset to the
end of the series (or something similar to this). It should be possible
to find a constellation like this that yields the same results but much
faster.